+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   @cosmjs/launchpad instead.
 - @cosmjs/cosmwasm: Export `JsonObject`, `ChangeAdminResult` and `WasmData`
   types as well as `isValidBuilder` and `parseWasmData` functions.
+- @cosmjs/cosmwasm: Add `CosmWasmClient.getTx` method for searching by ID and
+  remove such functionality from `CosmWasmClient.searchTx`.
 - @cosmjs/cosmwasm-stargate: Add new package for CosmWasm Stargate support.
 - @cosmjs/launchpad: Add `Secp256k1Wallet` to manage a single raw secp256k1
   keypair.